+++ packages/pangea-node-sdk/tests/integration/vault.test.ts
@@ -14,7 +14,7 @@ import {
   strToB64,
   TestEnvironment,
 } from "../../src/utils/utils.js";
-import { loadTestEnvironment } from "./utils.js";
+import { loadTestEnvironment, delay } from "./utils.js";
 
 const environment = loadTestEnvironment("vault", TestEnvironment.LIVE);
 const token = getTestToken(environment);
@@ -44,7 +44,7 @@ const KEY_ED25519 = {
 };
 
 const KEY_AES = {
-  algorithm: Vault.SymmetricAlgorithm.AES,
+  algorithm: Vault.SymmetricAlgorithm.AES128_CFB,
   key: "oILlp2FUPHWiaqFXl4/1ww==",
 };
 
@@ -919,3 +919,47 @@ it("export", async () => {
     exportedPlain.result.private_key
   );
 });
+
+afterAll(
+  async () => {
+    let last = undefined;
+    let count = 0;
+
+    while (count < 500) {
+      // List
+      const listResp = await vault.list({
+        filter: {
+          name__contains: ACTOR,
+        },
+        last: last,
+        size: 10,
+      });
+
+      listResp.result.items.forEach((element) => {
+        if (
+          element.id != undefined &&
+          element.type != "folder" &&
+          element.folder != "/service-tokens/"
+        ) {
+          vault
+            .delete(element.id)
+            .then(() => {
+              count++;
+            })
+            .catch((e) => {
+              console.log(e);
+            });
+        }
+      });
+
+      last = listResp.result.last;
+      if (listResp.result.items.length == 0) {
+        break;
+      }
+      await delay(5000);
+    }
+
+    console.log(`Deleted ${count} items`);
+  },
+  20 * 60 * 1000
+); // 20 minutes