+++ packages/react-mui-shared/src/components/PangeaFields/FieldsForm/fields/index.tsx
@@ -3,6 +3,10 @@ export { default as CheckboxField } from "./CheckboxField";
 export { default as DateTimeField } from "./DateTimeField";
 export { default as StringArrayField } from "./StringArrayField";
 export { default as JsonField } from "./JsonField";
-export { default as StringField } from "./StringField";
+export {
+  default as StringField,
+  UnControlledAutocompleteStringField,
+  UnControlledStringField,
+} from "./StringField";
 export { default as SwitchField } from "./SwitchField";
-export { default as SelectField } from "./SelectField";
+export { default as SelectField, UnControlledSelectField } from "./SelectField";