+++ packages/react-mui-shared/src/index.ts
@@ -17,6 +17,11 @@ export {
   LinedPangeaDataGrid,
 } from "./components/PangeaDataGrid";
 export {
+  AutocompleteValueOption,
+  ConditionalOption,
+} from "./components/ConditionalAutocomplete/types";
+export { OptionComponentProps } from "./components/ConditionalAutocomplete/AutocompleteOptionComponent";
+export {
   FieldsPreview,
   FieldsPreviewProps,
   FieldsPreviewSchema,
@@ -35,8 +40,11 @@ export {
   DateTimeField,
   StringArrayField,
   StringField,
+  UnControlledStringField,
+  UnControlledAutocompleteStringField,
   SwitchField,
   SelectField,
+  UnControlledSelectField,
 } from "./components/PangeaFields";
 export {
   default as PangeaModal,
@@ -61,6 +69,19 @@ export { default as JsonViewer } from "./components/JsonViewer";
 export { default as ConditionalAutocomplete } from "./components/ConditionalAutocomplete";
 export { default as PopoutCard } from "./components/PopoutCard";
 export { default as TimeRangeSelect } from "./components/TimeRangeSelect";
+export { default as RelativeDateRangeField } from "./components/RelativeDateRangeField";
+export {
+  getDisplayDateRange,
+  getRelativeDateRange,
+  getHalfRelativeDateRange,
+  compareRelativeDateRanges,
+  getRelativeDate,
+} from "./components/RelativeDateRangeField/utils";
+export {
+  TimeRangeString,
+  RelativeRange,
+  RelativeRangeObject,
+} from "./components/RelativeDateRangeField/types";
 export { default as CopyButton } from "./components/IconButtons/CopyButton";
 
 /* Exported utils */